Added flags for Fortran compiler; fixed warnings in unirxn, spectralfit.
All of the warnings reported by -Wall have been fixed for the Fortran files
in the rmg.unirxn module, including one variable that was used but not set.
A bounds error was also fixed.

Only the warnings reported by -Wall regarding variables that are used but not
set have been fixed for the Fortran files in the rmg.spectralfit module. There
are four such occurrences: three are fixed and one is (hopefully) okay.

Does this result in fewer segfaults? We'll find out.