Major reorganization of modules relating to spectral data.
There is now enough code devoted to treating spectral data that it deserves
its own package, which is rmg.spectral. As of now there are three modules in
the package:

* rmg.spectral.data - Contains the frequency database

* rmg.spectral.modes - Contains classes for the types of degrees of freedom

* rmg.spectral.fit - Contains functions for fitting spectral data

The Fortran code associated with the old rmg.spectralfit (including the
"Frankie" code) is also contained in rmg/spectral, and is compiled into the
rmg.spectral.fit module.