Fixed bugs in heat capacity expressions; added degeneracy to certain modes.
The new spectral data fitting function is now in working mode. Typos in the
heat capacity functions and derivatives in _modes.f90 and their use in
cases.f90 have been corrected.

A degeneracy attribute was added to the HarmonicOscillator and HinderedRotor
classes. This is especially useful for condensed representation of
pseudo-oscillators and pseudo-rotors as returned from the spectral data
fitting algorithm.