Added spectral fit case for fitting real oscillators and pseudo rotors.
This case is executed if not all modes can be fitted directly, but there are
few enough modes to fit that all oscillators can be fitted directly if one
pseudo-hindered rotor is fitted. Thus the criterion for using this case is
Nvib + 2 < len(Tlist) (for cases when Nvib + 2 * Nrot >= len(Tlist), of
course).