Added several checks and a few bugfixes to the spectral data estimation process.
Fixed bugs related to removing group modes when the number of frequency groups
found causes the molecule to be overspecified. This was causing negative
value of Nvib and/or Nrot to be passed to the fitting algorithm, which is
not good.

Prints a warning if maximum iterations are reached before the optimum solution
is found. Also prints a warning if the error in the fit is greater than a
certain root mean square value. These cases do not raise exceptions because
the heat capacity data used for the fitting may itself be the cause of the
poor quality of fit.

Added several places where SpectralFitExceptions will be raised due to
exceptional conditions being encountered. Hopefully these won't ever occur,
but if they do, they should cause a more graceful (and informative) failure.

Reduced the optimization tolerance from 1E-9 to 1E-5.