Added command-line option to specify the wall time.
RMG will now stop if it determines that it cannot complete another iteration
before the wall time is reached. The format for the wall time option is
HH:MM:SS, much like the PBS queue format. The wall time was made a
command-line option rather than an element in the input file because of the
typical use case of PBS queueing - this way you can specify the wall time to
both PBS and RMG in your PBS submission script.