Fixed warning in calculateBondSymmetryNumber() that occurred for diatomics.
The warning occurred when that function attempted to compare the ligands of
the atoms in the bond for equality to ensure the symmetry existed. For
diatomics there are no ligands, and so empty graphs were being passed to the
isomorphism check. The function now checks for diatomics first, and acts
accordingly to skip the ligand check for these species.