Added functionality for automatic parameter estimation to unirxn.
The user can now specify an input file to unirxn.py that only contains
structures for each species, and RMG will use its thermo and frequency
databases to estimate the needed parameters. This is not available for
reaction kinetics, so you must still specify the high-pressure-limit
kinetic parameters for the reactions.