Added error checking to chemical eigenvalue method in unirxn module.
The chemically-significant eigenvalue method requires two conditions on the
calculated eigenvalues in order for the method to be valid:

(1) The chemical and internal energy eigenvalues must be clearly separable.

(2) There should be a zero eigenvalue corresponding to the equilibrium
condition.

If either of these conditions is not met, then the output from the method is
not to be believed. Note that this signficantly restricts the range of
temperatures and pressures where the chemical eigenvalue method is viable.
(It's also rather slow!)