Improved chemical eigenvalues method: more straightforward k(T,P) evaluation.
My current understanding of the mathematics of this is that we want to create
a new master equation matrix M' with the internal energy eigenmodes removed.
This can be done with a formula like

M' = X * V * Xinv

where V contains only the chemically-significant eigenvalues. This means that
X is Nrows x Ncse, V is Ncse x Ncse, and Xinv is Ncse x Nrows. The k(T,P)
values can be extracted from M' in the same way that the kinf(T) values can
be extracted from M.

Also takes advantage of the fact that the eigenvector matrix corresponding to
a symmetric matrix is unitary to more rapidly generate Xinv.