Better version of reservoir cutoff grain selector for reservoir state method.
Using the ground-state energy to find the starting grain has proven to be
occasionally problematic because it might let a row of all zeros into the
matrix, causing it to be singular. A better test is to look at the density of
states, which is nonzero at all grains that should be included.