Repaired some unit tests to reflect changes to the codebase.
The goal of this commit was to eliminate ERROR results from the unit tests.
There are some unit tests remaining that FAIL.

* The way the log is initialized changed with the introduction of the rmg.log
module.

* The way connectivity information was stored in Atom objects had been
changed from a 3-tuple to three separate scalar variables.

* The way a type of thermo data to be generated is specified to
getThermoData() had changed from required_class to thermoClass.

* Importing rmg.species directly without previously importing rmg.thermo
causes a segfault for an unknown reason.

* canteraloadertest and simulationtest generate errors because of the switch
to using Cantera as the solver.