Quieten down unit tests a little.
I've removed simulationtest and canteraloadertest from the test.py suite.

canteraloadertest was used to test the cantera_loader.py stuff which allowed us
to read in cantera and chemkin models. This is rather broken, but may be worth
repairing in some form in the future.

simulationtest requires cantera_loader.py but used to check our internal
reactor model and ode solver against a cantera simulation, but as we're now
using cantera this is a little perverse. This test is probably not worth
reviving.