Changed stoichiometry matrix to use species and reaction id for indexing.
The new stoichiometry matrix is much larger, but is also sparse, so the memory
use scales as N rather than N^2. The tradeoff should be a large increase in
speed, since Species.id is an O(1) attribute, while list.index() is an O(N)
function. The matrix is still completely regenerated prior to each simulation.