Changed definition of characteristic flux to only consider core reactions.
The previous implementation counted core and edge reactions when evaluating
the root mean square of the core species fluxes. For early models with no
core reactions, this causes the characteristic flux to be nearly equal to the
maximum flux to the edge, when we were expecting a characteristic flux of
zero.

Also added a new function BatchReactor.getCharacteristicFlux() for evaluating
the characteristic flux at each time point in the simulation. The current
implementation of this function is not particularly fast, as it adds a
second call to getReactionRates().