A new way of doing pressure dependent rates in Cantera
ctml_writer has global variables _temperature and _pressure
these are used to evaluate a pressure dependent rate coefficient 
before converting it into a rate constant for writing the xml

system/batch updates these _temperature and _pressure variables before calling write(). (It also now creates the initial_state based on self.initialTemperature and self.initialPresure, so the cti file should be internally consistent)

PDepReaction now has its own toCantera method.
This calls the super class Reaction's method of the same name,
then replaces the _kf (which was a ctml_writer.Arrhenius) with a 
ctml_writer.PdepRate, which we initialize with self.getRateConstant

ctml_writer.PdepRate is a new class of rate coefficient in ctml_writer.
Instead of giving it some parameters from which it can calculate k, 
you give it a *function* of T and P. When asked to build itself into xml,
it calls this function using _temperature and _pressure, creates an 
Arrhenius rate with the equivalent k, and builds that. 

The one thing I'm not sure of is the units of P assumed by and the units of
k returned by PDepReaction.getRateConstant(T,P) and by Cantera's Arrhenius A.