Better handling of negative activation energies in ILT method.
The inverse Laplace transform method is used for k(T) -> k(E) conversion. The
method requires that Ea > 0. Occasionally this was not the case due to bad
kinetic parameter estimation. Before we would just set Ea to zero with no
correction to A; this worked okay because the magnitude of the negative Ea
was often small. However, it is more rigorous to correct A as well.