Moved rmg.unirxn.states to rmg.spectral.states.
The density of states calculation belongs with the spectral module, not with
the unirxn module. In particular, there were two Fortran implementations of
a few density of states functions, one of which was out of date. This move
eliminates those and ensures that every part of the code is using the same,
up-to-date functions.