Fixed some "Child is not a child" instances
Addresses parts of Issue #30

Some of the chemgraph defintions in the Other_Dictionary.txt and
Ring_Dictionary.txt files were changes to address the "Child is not
a child" issue.  In most instances, the starred (*) central atom was
incorrect.

The remaining warning messages emanate from Abraham_Dictionary.txt and
Unifac_Dictionary.txt