Added auto temperature and pressure range option for pdep networks.
These are specified in the input file using the <temperatureRange> and
<pressureRange> elements instead of the <temperatures> and <pressures>
elements. For each of the former, you specify attributes units, Tmin, Tmax,
and number.

For a Chebyshev interpolation model the temperatures and pressures will be
spaced on a Gauss-Chebyshev grid (similar to Chemkin). For all other models
the temperatures and pressured will be evenly distributed on 1/T and log10(P)
scales, respectively.