isSpecifecCaseOf(self,other) added to many chem Classes
Returns True if self is equal to other or a more specific case of other.
Returns False if self is more general than other or some of it is not included in other or they are mutually exclusive.

Added to 
 AtomType
 ElectronState
 BondType
 Atom
 Bond

NB. these have not yet been added to the cython .pxd file