Graph matching uses isSpecificCaseOf() for subgraph matches
When subgraph=True is passed to the graph isomorphism functions
it uses the isSpecificCaseOf methods on the atoms and bonds.

I believe VF2_isomorphism(graph1,graph2,...subgraph=True) 
is like graph1.isSpecificCaseOf(graph2)

I think the behaviour when subgraph=False (i.e. it's looking
for a specific match) is probably incorrect, in that the
'equivalent' methods are currently too generous. But I'm not 
sure when it is used - perhaps we assume the current behaviour
so 'fixing' it would break something. Worth checking.