Added setting for controlling restart file saving to input file.
As the generated mechanisms become large, the restart file saving begins to
occupy an increasingly signficant amount of the execution time. Adding this
option gives the user the ability to mitigate this as they wish. The new
option has the form

<saveRestart frequency="daily" iterations="10"/>

You can optionally specify the frequency (in time - one of "always,
"hourly", "daily", "weekly", or "monthly") and/or the number of iterations
after which a restart save should be forced. To turn off restart file saving,
omit this element. If <saveRestart/> is found with no options, it will save
the restart file after every RMG iteration.

The location of the restart file is not provided as an option on purpose, so
that it can be easily found when using the -r option.