Fixed two minor bugs in pressure dependence code.
The first deals with the equilibrium ratios passed to the eigenvalue method.
The prior implementation was incorrect because it did not account for the
difference between unimolecular and bimolecular configurations. This has now
been corrected.

The second deals with the halfbandwidth of the collision matrices. The
original formula had a tolerance of 1e-16, which is far too high. The new
version uses 1e-6, which should speed up the computation a bit.