chem.Atom and chem.Bond now inherit from graph.Vertex and graph.Edge.
This refactoring allows us to store the connectivity and sort_label helper
attributes in Vertex rather than Atom, which makes more sense because they
are structural properties rather than semantic ones. Also, this refactoring
means that the graph module is now independent, while the chem module now
depends on graph; prior to this change it was the other way.