New procedure for reaction generation within a reaction family.
The new procedure is as follows:

* All of the possible reactions involving one or two reactants within that
family are generated and Reaction objects are created for them. Duplicate
checking is not yet performed, nor is kinetics estimation.

* The list of generated reactions is checked for duplicates. If found, the
duplicate is deleted and the original has its multiplier incremented. This
should give a much more accurate estimate of the reaction path degeneracy.

* Finally, the remaining generated reactions are placed into the global list
of reactions and their kinetics are generated. Duplicate checking with the
global list is performed here, but the multipliers are not changed.

This procedure does not yet work for A + A --> products reactions.