Reduced the number of times Structure.copy() is called.
ReactionFamily.applyRecipe() originally called copy as the combined
reactantStructure was being created *and* after applying the recipe to create
the combined productStructure. Only one of these is required, and only then
for the case A + A --> products when using the same structure for both A
molecules. This should save a bit of time and perhaps some memory as well.

Structure.copy() also now optionally returns a mapping of the copy to the
original.