Reaction is put in forward direction according to reaction family (or dHrxn<0)
Previously makeNewReaction would take the reaction in the direction it was given,
try to figure out kinetics in both directions, and use whatever succeded to 
determine which was 'forward'.

Now, whether or not a family is forward or reverse is stored in the family
when first read from the database. We use this information to put reactions
into the forward direction before determining their kinetics. If both directions
are marked as forward, (eg. H abstraction) then the exothermic direction is used 
as forward.

Behaviour for H abstraction is better. Net behaviour for other families is the same,
except that now the call to checkForExistingReaction is always in the same
direction!