Use global reactionDict of short-lists rather than long reactionList
reactionDict[family][reactant1][reactant2] is a list of reactions.
If you combine all these short-lists into one long list you'd get back 
the old reactionList. The benefit of splitting it up is that instead 
of searching through the entire list to find a reaction, we look up the 
shortlist, and only need to search through that!

NB. reactants are always reactants because reactions are now always made 
in the same forward direction, EXCEPT for seed reactions, in which case
you have to check for ['seed'][product1][product2] too.