Fixed a few small bugs in the density of states evaluation algorithm.
The Beyer-Swinehart function was not adjusted to expect the energies in J/mol
when the other functions were, so it was incorrectly convolving in the
vibrational modes. This has been corrected.

A slight modification to the cellipk() function - used for evaluating the
complete elliptical integral of the first kind - was made to exactly match
the algorithm used in MATLAB.