Some tweaking of the spectral mode fitting algorithm.
The objective is to reduce the number of instances that the "pseudo"
case is used because we feel that fitting degeneracies causes extra
numerical difficulties on top of the high amount of nonlinearity.
One way to do this is to increase the number of data points used,
which lets us increase the number of optimization variables.