Added primary frequency database.
This enables us to manually specify the vibrational and hindered rotor modes
for species that we have detailed information for. This is particularly useful
for certain small molecule cases that the group frequency method would find
to be overspecified; these are what are currently in the primary database.
However, other species can also be added.