Added license file and readme.
There are a lot of chemistry toolkits out there, some of which we will
incorporate and others that we will compete against. We want to encourage
wide use of our code, and the real value comes more from what you do with it
than the codebase itself. Thus the MIT license has been chosen.