Added modules for numerical constants and for exceptions.
The constants module can be compiled via Cython, but I don't think it provides
as much of a speed boost as it ought to. The current version of Cython doesn't
seem to like module-level variables at the moment.

The exception module is a place to collect all ChemPy-related exceptions. This
module is not compiled via Cython, as we don't expect the circumstances when
an exception would be thrown to be time critical.