Added thermo and kinetics methods to Species and Reaction objects.
These methods allow for easy evaluation of the thermodynamics of Species
objects and the thermodynamics and kinetics of Reaction objects. Some effort
has been put in to Cythonize these methods as much as possible -- by declaring
the type of all parameters, return values, and local variables -- because we
anticipate calling them frequently within time-critical code. There may yet
be more Cythonizing that can be done to these methods, however.