Added graph module, containing an implementation of a graph data structure.
The graph data structure is useful when representing chemical structures in
memory: atoms correspond to vertices and bonds correspond to edges.
Accordingly, there are three useful extension types in this module: the
Vertex, Edge, and Graph classes.

This file was copied intact from the RMG Py sources, with only a few small
modifications.