Fixed misunderstanding in pressure dependent network expansion process.
The key to getting proper error control using the so-called "activated species
algorithm" developed by D. M. Matheu is to treat the unexplored unimolecular
isomers as irreversible product channels. The MSC and RS methods were already
able to handle this; this commit enables proper input to be passed to these
methods, i.e. that unexplored unimolecular isomers are treated as irreversible
product channels. This has the added benefits of speeding up the pdep
calculation and reducing the likelihood of the RS method failing due to
stiffness because the master equation matrix sizes of the partial networks is
reduced.

Note that the parameter list to Network.applyApproximateMethod() has changed
to require the passing of nIsom and nReac. This method also assumes that the
isomers have been sorted such that the first nIsom are the unimolecular
isomers, the next nReac are the bimolecular reactant channels, and the rest
are unimolecular or bimolecular irreversible product channels.