Improved version of HarmonicOscillator class for representing vibrations.
In addition to the changes made parallel to those of the Translation and
RigidRotor classes in the previous commits, there is a significant new change
with HarmonicOscillator objects: one HarmonicOscillator object now can
represent all vibrations in a molecule. This makes the calculation of the
thermo more efficient and allows for placing the Beyer-Swinehart algorithm
in the getDensityOfStates() method.