Improved version of HinderedRotor class for modeling hindered internal torsion.
After much thought I decided to only implement the simple potential function
in this version of the class. There may be demand for a Fourier series
potential in the future; I am not sure if this will be its own class or will
be built into the same HinderedRotor class.

I also decided that each HinderedRotor object will only represent a single 1D
rotor because each rotor consists of multiple parameters.

The equations used for the HinderedRotor should be closer to correct than they
were, but could use a double-check. I don't have explicit equations for the
enthalpy and entropy of a hindered rotor, so these are currently evaluated
numerically from the partition function.