Lots of tidying up of the states module.
Cleaned up the docstrings of almost all of the methods to make sure they
referred to all parameters in the parameter list.

Changed heat capacity, enthalpy, and entropy functions to return these
values in dimensionless units. Would be nice to also do this for the
density of states functions, but I'm not sure how to do so.

Added a bunch of cython local variable declarations to hopefully give the
methods of the various classes a speed boost when compiled.