Improved version of StatesModel class.
StatesModel is designed to contain the output from a quantum chemistry
calculation. In particular, it contains the output from a calculation similar
to running a "Freq" job in Gaussian. Its methods use this information to
calculate the same parameters that can be calculated for each individual
degree of freedom: partition function, density of states, heat capacity,
enthalpy, and entropy.