PDep: MSC, RS, and CSE methods now return population distribution vectors.
Strictly speaking, each of the approximate methods returns a k(T,P) matrix
and a set of time-independent vectors that, when combined with the
time-dependent concentrations x(t) and yA(t)*yB, allow for construction of
the approximate p(E,t). For RMG we're only really interested in the total
concentration for RMG, but others may find a use for the other data.

Each of the approximate methods reduces the variables from p(E, t) to x(t) by
assuming the variables are separable:

    p(E, t) = x(t) p0(E)

In order to reconstruct the p(E, t) data, we need the p0(E) data. Thus this
data is now returned by each of the MSC, RS, and CSE methods, and therefore
also returned from applyApproximateMethod().

Also removed a rescaling from the RS method, which interfered with the proper
calculation of the population distribution vectors.