PDep: Can now calculate k(T,P) values from population distribution vectors.
A new private method __calculateRatesFromPopulationVectors() was created for
this purpose. It is private because it is designed to be used from within
applyApproximateMethod() only.

This is primarily to demonstrate that the common formalism does in fact work;
I don't think this is a good way to do it in production code because it is
not as optimized as it is for the individual methods themselves. For this
reason, it remains commented out.

Tested with both nProd = 0 and nProd > 0, as well as with nIsom > 1 and
nReac > 1.