Added geometry molecule for working with molecular geometries.
The initial impetus was to write functions for calculating moments of inertia,
both of an entire molecular configuration and for individual internal hindered
rotors. Functions for both have been written as methods of the Geometry class.

I am not sure yet how exactly this will fit into the storing of structure
information. My inclination is to keep it mostly separate so that it doesn't
take up much memory for other uses that don't utilize this functionality.