Significant refactoring of graph module.
* The Graph class no longer inherits from dict. Instead, it provides a list
for vertices and a dict of dicts for edges, both as attributes of the class.
Much of the refactoring involved changing methods around to reflect this
change.

* Removed several places where a chem.Atom was used in place of a Vertex and
a chem.Bond in place of an Edge. Since I figured out how to allow Atom and
Bond to inherit from Vertex and Edge, the graph module can be made to be
standalone.

* Removed sorting_label as an attribute. If the function that calculates this
value proves to be a bottleneck, it can be restored.

* Removed cycle methods from graph class (for now).