Added molecule module for working with molecular structures.
As currently implement, the molecule module contains four classes: Atom, Bond,
ChemGraph, and Molecule. The former tree extend the Vertex, Edge, and Graph
classes in the graph module with attributes and methods that pertain to atoms,
bonds, and molecular structures. The Molecule class represents a single
molecular structure as a list of its resonance forms. Each resonance form is
a ChemGraph object and corresponds to a valid Lewis structure.

The easiest way to create a molecule object is from its SMILES, InChI, or CML
representation. This functionality is provided by OpenBabel, and thus to use
it you must have OpenBabel (and its Python wrapper pybel) installed.

Also added getElement() function to element module to look up elements by
symbol or atomic number.