ThermoModel classes now use numpy arrays in parameters and return values.
This makes them consistent with the classes in the states module. The use of
numpy arrays throughout gives the end user a nice speed boost even when in
pure Python mode. It also makes sense, as this project is named *Py, to use
NumPy everywhere.

Also note that the heat capacity, enthalpy, entropy, and free energy functions
now return *dimensionless* values (i.e. scaled by R or RT), again to be
consistent with the states module classes.