Added functions for fitting WilhoitModel to Cp(T) vs. T data.
The default is to use an optimization routine to also optimize the B value.
The others are optimized at a single value of B using a linear least-squares
procedure.

This commit exposes an unexpected bug in Cython 0.12.1, wherein having
multiple derived classes with the same base class causes the last derived
class's methods to trample over those of the earlier derived classes. Thus
a few parts of the thermo module don't work as expected when compiled.