Added functionality for transition state theory calculations (with tunneling).
A new class TransitionState is available from the species module for working
with transition states. Unlike full Species objects, transition states don't
have thermo models or identifiers; instead, they have an attribute for the
negative frequency (needed for the tunneling correction functions).

The transition state theory functions are methods of Reaction objects. Their
implementation follows very closely to that of CanTherm, but with an extra
wave of comments and better Pythonic code structure.

Also added a bunch of Cython declarations to hopefully speed things up a bit.
More work in this area could be done if needed.