Added first version of introduction to Theory Guide.
The introduction is designed to briefly describe the physical phenomena
involved with unimolecular reactions and give a brief history of how our
understanding has evolved with time.

Note that a bunch of literature citations have been added. These should
generally be placed at the end of each section. Also, the DOI or some other
form of permalink should be provided if at all possible with these references
to make it as easy as possible for the interested reader to follow up and
learn more.