Added master equation section to theory manual.
This is a rather long-winded discussion of what goes into the master equation.
Unlike the manuscript, I include discussion towards both the 2D and 1D
versions of the master equation, as I am thinking of driving the codebase in
this direction.

A few lines were added to the LaTeX header to declare a few commands that my
LaTeX equations use extensively.