Moved E0 from StatesModel to Species; added loadEnergy() to GaussianLog.
The ground-state energy E0 doesn't really go with the StatesModel class, as
the latter is designed to work with energies referenced such that the zero of
energy is the ground state energy. Furthermore, there are cases where you
might know E0 but not the modes (e.g. a Gaussian "Opt" or "Energy" job), and
don't want to create a StatesModel object just to store E0.

Also fixed density of states and partition function calculation methods to
include the spin multiplicity.

These changes required refactoring a few unit tests. All unit tests continue
to pass.