Added primary script for running jobs, with command-line argument parsing.
The argparse package is used for parsing command-line arguments. Users of
Python versions prior to 2.7 will need to install this themselves; users of
Python 2.7 or later will find it already in the Python standard library.

Currently the script parses the command-line arguments, initializes the
logging system (to console only for now), and prints an initiation timestamp,
identifying header, and termination timestamp.