Created measure package; added first version of input file reading module.
Input files are designed to have a Python-like syntax to make them easy to
read in Python: just connect appropriate identifiers to functions and classes!
Preliminary support for reading species and reaction data from the file has
been added, although the format is far from final.

The input module is also somewhat ugly due to the in-progress nature of the
input file format. More comments and docstrings are a must!