Added ability to specify temperatures, pressures, and energies in input file.
The temperatures and pressures can be specified either as lists of values to
use or by specifying a range and a number of points. The energies must be
specified using a range for now.

Changed readInput() to return the temperature, pressure, and energy parameters.
However, if the input is invalid for whatever reason, all returned values will
be None. This is done to allow the program to exit gracefully (with a
termination timestamp) when invalid input is discovered.