Added functionality for automatically determining the energy grains.
This functionality is codified in the method Network.autoGenerateEnergyGrains(),
which is automatically called if you do not specify an energy range in the
input file. The method uses equilibrium distributions to determine a suitable
maximum energy at which to truncate. My experience has been that the truncation
is plenty high in nearly all cases.