Created collision module for working with collision models.
Ultimately the goal is to allow for a bunch of collision models, and leave it
to the end user to provide the parameters needed for the more advanced models.
At the moment only the single exponential down model is implemented because it
is the simplest and most commonly used.

Methods for calculating the collision frequency and MSC collision efficiency
have also been provided.