Added reaction module for working with microcanonical rate coefficients.
The primary purpose of this module is to determine k(E) values for each
reaction. It will automatically choose the most accurate method that it can
for the amount and type of information provided (currently either RRKM theory
or inverse Laplace transform).

The inverse Laplace transform method is currently limited to Arrhenius kinetics
where n >= 0. A temperature-independent method must be developed for n < 0
and for non-Arrhenius kinetics (numerical inversion perhaps?).