Minor changes to translation entropy function and associated unit tests.
The entropy function for translation was only correct for 3D motion; it is now
correct for other numbers of dimensions as well.

The unit tests comparing the states from ethylene and oxygen Gaussian
calculations are improved in that they allow direct comparison of the entropy
values calculated by us to those in the Gaussian log file. This confirms that
the new version is correct.