Added getDensityOfStatesILT() method to StatesModel.
This method allows you to approximately determine the density of states by
inverting the inverse Laplace transform via the method of steepest descents
(as described by Forst). This could be useful e.g. if you have a lot of
hindered rotors and want a more accurate density of states. It is definitely
not the fastest procedure around, which detracts from its usefulness.