Revamped Translation class with better handling of molecular volume.
The translation functions I've been using implicitly assume 3D translation of
an ideal gas. This means the volume is RT/P, which implies a temperature
dependence. For the purposes of this class a pressure of 1 bar is assumed; you
can adjust to other pressures as desired.

Updated unit tests accordingly.