Added multiple PES drawing targets based on extension; now drawing to PDF.
Cairo gives you the option of drawing PDF, PS, or SVG files. Of these, PDF is
most likely to be immediately readable by end users, so we will go ahead and
use that as the default.

The output file (PES.pdf) is saved to the directory containing the input file,
not the current working directory. Future output will also be saved to that
location.