Added first version of molecule drawing module in new ext package.
The ext package is for extra functionality that is outside the core but still
belongs with this library. Molecule drawing is a good example of this, as it
is not required (and depends on Cairo if you want to use it) but doesn't
deserve a standalone project of its own. I didn't want to put it into the
molecule module because it requires a fairly significant amount of code.