Can now determine coordinates for straight-chain functional groups.
The key here is the recursive function processFunctionalGroup(), which allows
for easy handling of branching functional groups. The bonds are placed such
that they radiate away from the origin -- which should be set to the center of
the backbone at this point -- to reduce likelihood of overlap.

Also made a few minor refinements to the rendering algorithm.