Lots of improvements to the way symbols are drawn in molecule renderings.
The algorithm will automatically detect whether the central heavy atom should
come first or last, layout the characters of the label (including subscripts
for numbers), and colorize based on the type of the central heavy atom.

The background for these was changed to a rounded rectangle to give better-
looking diagonal bonds.