Some improvements to molecule drawing as a result of checking small molecules.
Perhaps most importantly, the algorithm will actually *handle* small molecules
without crashing or returning an empty picture.

Also refined the procedure by which radicals and charges are placed, as the
previous one had a few inconsistencies. (These were especially noticed because
the height of "-" is much less than that of "+" and alphanumeric characters.)