Now determines approximate size of molecule drawings.
It finds the atoms on each edge of the bounding box and uses their coordinates
as the first guess. It then uses a dummy Cairo surface/context pair to get
the width of the symbols being used for these atoms (a slight overestimate if
subscripts are present, but okay for now) and enlarges the box accordingly.
Finally, a small amount of padding is added on to all sides.