First version of ring drawing algorithm in molecule drawing code.
The general idea is to find all of the independent ring groups -- which are
separated by one or more non-cyclic single bonds -- and generate the
coordinates for each ring group independently. The largest of these groups
will form the molecular backbone, while the others will be translated and
rotated appropriately as the functional groups are positioned.

This code depends on a good method for identifying the ring groups, which we
don't have yet. I also haven't written the part than handles rings in
functional groups.