Can now handle multiple ring groups in molecule drawings.
The largest ring group becomes the molecular backbone, while the others are
drawn as part of the functional groups of the backbone.

A new function generateNeighborCoordinates() was created to be used when
processing the immediate neighbors of the backbone (either cyclic or acyclic)
and cycles found within functional groups. This function can now handle both
cases where the bonds are arranged evenly around the molecule, as well as
cases where they are not (e.g. when a cycle is present).