Significant refactoring and commenting of molecule drawing functions.
At this point the molecule drawing algorithm is complete enough that it works
for a wide range of molecules. I'm sure there will be a few bugs to fix, and
that one or more incomplete features might need adding, but this version is
far enough along to begin to be useful.