Removed dependence on quantities package; updated package dependencies lists.
The quantities package is neat and can be useful, but is also rather slow to
load and doesn't seem to work as well on Python 2.5. At this moment we were
only using it to get values of physical constants, which we can do just as well
manually.

Cairo has been added to the list of dependencies required to utilize the full
functionality.