Created pattern module; first version of AtomPattern class.
The pattern module provides functionality for molecular substructure pattern
matching. The module will be laid out similarly to the molecule module, with
equivalent classes to Atom, Bond, and ChemGraph for representing molecular
patterns. We need new classes because we want to allow for wildcard atom and
bond types.

The AtomPattern class provides the ability to represent wildcard atoms in
molecular patterns. Two sets of methods have currently been implemented: one
for changing the atom types as a result of applying a reaction recipe action
(AtomPattern.applyAction() et al) and another for semantic comparisons
(AtomPattern.equivalent() and AtomPattern.isSpecificCaseOf()).

A note on imports: we will have the molecule module import the pattern module,
not vice versa.